requirements.txt
tk
Pillow
pandas
pyinstaller